+++ lib/src/constants.dart
@@ -1,0 +1,5 @@
+const int vnTimezone = 7;
+const int daysInYear = 365;
+const int yearIndex = 2;
+const int monthIndex = 1;
+const int dayIndex = 0;